Moved both polyfills into the same folder.
