+++ hot/polyfillEnvironment.js
